% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_plot.R
\name{barcode_plot}
\alias{barcode_plot}
\title{barcode_plot}
\usage{
barcode_plot(pathway, gsea_pathways, gsea_vector, gsea_results)
}
\arguments{
\item{pathway}{the chosen pathway to plot.}

\item{gsea_pathways}{your list of pathways used for GSEA.}

\item{gsea_vector}{the ranked vector you used for GSEA with your human gene names and logFC.}

\item{gsea_results}{your GSEA results dataframe.}
}
\value{
A nice looking barcode plot for your chosen pathway with gene ranks.
}
\description{
Create GSEA barcode plot from a chosen pathway.
}
