% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Rsq.R
\name{calculate_Rsq}
\alias{calculate_Rsq}
\title{calculate_Rsq}
\usage{
calculate_Rsq(dataframe, var_1, var_2)
}
\arguments{
\item{dataframe}{dataframe with the values you're using for linear regression.}

\item{var_1}{one of the variables you're using for linear regression inputted as string.}

\item{var_2}{the other variable you're using for linear regression inputted as string.}
}
\value{
Rsq value.
}
\description{
calculate Rsq value from linear regression.
}
