% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcycle_bar.R
\name{cellcycle_bar}
\alias{cellcycle_bar}
\title{cellcycle_bar}
\usage{
cellcycle_bar(seurat_obj, group_by)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{group_by}{the ident from your Seurat object that you want to group the cells by.}
}
\value{
bar graph.
}
\description{
create bar graph of \% of cells in each cell cycle phase across different Seurat object idents.
}
