% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_FindMarkers_human.R
\name{convert_FindMarkers_human}
\alias{convert_FindMarkers_human}
\title{convert_FindMarkers_human}
\usage{
convert_FindMarkers_human(
  genelist,
  remove_ribosomal_genes = T,
  create_ranked_vector = T
)
}
\arguments{
\item{genelist}{FindMarkers output dataframe. Needs to have a column with the gene names called "gene" to proceed with this function.}

\item{remove_ribosomal_genes}{default = T, remove ribosomal genes from FindMarkers results?}

\item{create_ranked_vector}{default = T, create ranked vector to use with fgsea?}
}
\value{
ranked vector or dataframe with converted human genes.
}
\description{
convert the results of FindMarkers (dataframe) to human genes and create ranked vector for GSEA. * Improved version!
}
