% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_FindMarkers_human_old.R
\name{convert_FindMarkers_human_old}
\alias{convert_FindMarkers_human_old}
\title{convert_FindMarkers_human_old}
\usage{
convert_FindMarkers_human_old(
  fish_genelist,
  create_ranked_vector = T,
  remove_ribosomal_genes = T
)
}
\arguments{
\item{fish_genelist}{FindMarkers results dataframe.}

\item{create_ranked_vector}{if T, will convert the output into a ranked vector with human genes as name and avg_logFC as values, to use for GSEA.}

\item{remove_ribosomal_genes}{if T, will remove anything starting with rps or rpl from the gene list}
}
\value{
if create_ranked_vector = T, a ranked vector with human genes as name and avg_logFC as values, to use for GSEA. if F, a dataframe.
}
\description{
convert the output of the Seurat function FindMarkers from fish genes to human.
}
