% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_motif_targets_human.R
\name{convert_motif_targets_human}
\alias{convert_motif_targets_human}
\title{convert_motif_targets_human}
\usage{
convert_motif_targets_human(
  motif_genes,
  expression_data,
  create_ranked_vector = T
)
}
\arguments{
\item{motif_genes}{character vector with the list of genes enriched in your motif of interest.}

\item{expression_data}{FindMarkers results used to create original genelist for HOMER.}

\item{create_ranked_vector}{default = T, create ranked vector to use with fgsea?}
}
\value{
ranked vector or dataframe with converted human genes.
}
\description{
convert the genelist from HOMER, i.e genes enriched in a particular motif, to a ranked vector for GSEA, using the original FindMarkers results and adding human gene names. * Improved version!
}
