% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_motif_targets_human_old.R
\name{convert_motif_targets_human_old}
\alias{convert_motif_targets_human_old}
\title{convert_motif_targets_human_old}
\usage{
convert_motif_targets_human_old(
  motif_genes,
  expression_data,
  create_ranked_vector = T
)
}
\arguments{
\item{motif_genes}{character vector of the motif target genes (fish).}

\item{expression_data}{dataframe of the corresponding expression data for your genelist (usually output of FindMarkers). NOTE: the function assumes the gene names are in column 2 of this dataframe - if you're getting an error, check that first.}

\item{create_ranked_vector}{if T, will create a ranked vector of the human orthologs and expression data - to be used for GSEA.}
}
\value{
if create_ranked_vector = T, will return a ranked vector to be used for GSEA. if F, a dataframe.
}
\description{
convert HOMER motif target genes to human and add corresponding data - intended for cases when you want to perform GSEA on the HOMER motif target genes.
}
