% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_human_list.R
\name{convert_to_human_list}
\alias{convert_to_human_list}
\title{convert_to_human_list}
\usage{
convert_to_human_list(genelist)
}
\arguments{
\item{genelist}{list of fish genes as character vector.}
}
\value{
a character vector of the human orthologs of the given fish genes.
}
\description{
convert a fish genelist (character vector) to human, using DIOPT scores to remove duplicated genes. Intended to be used in cases where you don't have corresponding expression data for the genelist.
}
