% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_term_bar.R
\name{go_term_bar}
\alias{go_term_bar}
\title{go_term_bar}
\usage{
go_term_bar(
  gsea_results,
  n_terms = 10,
  metric = "NES",
  fill = T,
  change_labels = T,
  pval_fill = T
)
}
\arguments{
\item{gsea_results}{dataframe with GSEA results.}

\item{n_terms}{number of top GO terms to plot.}

\item{metric}{must be "NES" or "pval".}

\item{fill}{if T, will fill bars with gradient colours, if F, will fill bars with grey only.}

\item{change_labels}{to change pathway names to lowercase and remove "GO:" from the pathway names.}

\item{pval_fill}{if T, will colour bars based on pval.}
}
\value{
bar graph.
}
\description{
nice looking bar graph of top GO terms from GSEA by pval or NES.
}
