% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_dim_plot.R
\name{nice_dim_plot}
\alias{nice_dim_plot}
\title{nice_dim_plot}
\usage{
nice_dim_plot(
  seurat_obj,
  group_by = "seurat_clusters",
  cols = NULL,
  pt_size = 1.3,
  label = T,
  show_legend = F,
  reduction = "umap",
  dims_plot = 1:2
)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{group_by}{what to colour the points by, usually a column in the Seurat object metadata.}

\item{cols}{optional: colours to label the grouped points by.}

\item{pt_size}{point size.}

\item{label}{label groups with text on plot?}

\item{show_legend}{if T, will include the legend but won't label points w/text on plot.}

\item{reduction}{either "umap" or "pca".}

\item{dims_plot}{dimensions to plot.}
}
\value{
UMAP or PCA plot.
}
\description{
nicer looking version of the Seurat function DimPlot.
}
