% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_feature_plot.R
\name{nice_feature_plot}
\alias{nice_feature_plot}
\title{nice_feature_plot}
\usage{
nice_feature_plot(
  seurat_obj,
  features,
  pt.size = 1.3,
  n_col = NULL,
  reduction = "umap",
  dims_plot = 1:2,
  diverging_cmap = F,
  scale_data = F,
  cutoffs = NA,
  order = F
)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{features}{genes to plot.}

\item{pt.size}{size of plotted points.}

\item{n_col}{if plotting multiple genes, number of columns in final plot_grid object.}

\item{reduction}{one of "umap" or "pca".}

\item{dims_plot}{number of dimensions to plot.}

\item{diverging_cmap}{if T, will fill points with diverging colours between blue and red. otherwise, will use a linear colourmap.}

\item{scale_data}{plot Z-scored data?}

\item{cutoffs}{expression level cutoffs for plotting.}

\item{order}{plot cells in order of expression?}
}
\value{
UMAP or PCA plot.
}
\description{
nicer looking version of the Seurat function FeaturePlot.
}
