% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_scatter_plot.R
\name{nice_scatter_plot}
\alias{nice_scatter_plot}
\title{nice_scatter_plot}
\usage{
nice_scatter_plot(df, x_var, y_var, fit_line = T, calculate_rsq = T)
}
\arguments{
\item{df}{dataframe with variables to plot.}

\item{x_var}{variable to be plotted on X axis inputted as string, should be a column in df.}

\item{y_var}{variable to be plotted on Y axis inputted as string, should be a column in df.}

\item{fit_line}{set to T to perform linear regression and fit a line.}

\item{calculate_rsq}{set to T to calculate Rsq of line fitting.}
}
\value{
scatter plot.
}
\description{
create nice looking scatter plot with optional linear regression/line fitting.
}
