% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_scatter_plot_w_outliers.R
\name{nice_scatter_plot_w_outliers}
\alias{nice_scatter_plot_w_outliers}
\title{nice_scatter_plot_w_outliers}
\usage{
nice_scatter_plot_w_outliers(
  df,
  x_var,
  y_var,
  variable_to_label,
  low_outlier_cutoff = -0.5,
  high_outlier_cutoff = 0.5,
  fit_line = F,
  calculate_rsq = T,
  calculate_R = T,
  label_outliers = T
)
}
\arguments{
\item{df}{dataframe with data to plot.}

\item{x_var}{name of X variable column in df as string.}

\item{y_var}{name of Y variable column in df as string.}

\item{variable_to_label}{column in df with the labels you want on the plot for outliers.}

\item{low_outlier_cutoff}{bottom residual cutoff to select outliers, default -0.5}

\item{high_outlier_cutoff}{top residual cutoff to select outliers, default 0.5}

\item{fit_line}{plot a fitted line on the plot?}

\item{calculate_rsq}{calculate Rsq?}

\item{calculate_R}{calculate R?}

\item{label_outliers}{label outliers on plot?}
}
\value{
scatter plot.
}
\description{
nicer looking scatter plot with outlier points labelled.
}
