% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_spatial_feature_plot.R
\name{nice_spatial_feature_plot}
\alias{nice_spatial_feature_plot}
\title{nice_spatial_feature_plot}
\usage{
nice_spatial_feature_plot(
  seurat_obj,
  features,
  im_alpha = 0,
  pt_size = 1.4,
  stroke = 0,
  new_cmap = T,
  cmap = "inferno",
  n_col = NULL,
  diverging_cmap = T,
  alpha = T,
  cols = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{features}{genes to plot.}

\item{im_alpha}{set to 1 to plot the tissue image, 0 otherwise.}

\item{stroke}{linewidth to outline plotted points in black (default = no outline).}

\item{new_cmap}{if T, change the colourmap from Seurat default.}

\item{cmap}{colourmap to use if new_cmap = T.}

\item{n_col}{if plotting multiple genes, number of columns in final plot_grid object.}

\item{diverging_cmap}{if T, will fill points with diverging colours between blue and red. otherwise, will use a linear colourmap.}

\item{alpha}{if T, will fill points such that lower expression values are more transparent.}

\item{cols}{optional: gradient vector of colours to use for plotting.}

\item{pt.size}{size of plotted points.}
}
\value{
SpatialFeaturePlot.
}
\description{
nicer looking version of the Seurat function SpatialFeaturePlot.
}
