% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_violin_plot.R
\name{nice_violin_plot}
\alias{nice_violin_plot}
\title{nice_violin_plot}
\usage{
nice_violin_plot(
  seurat_obj,
  features,
  group_by = NULL,
  cols = NULL,
  pt.size = 0.3,
  sort = T,
  n_col = NULL,
  plot_hline = T
)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{features}{genes to plot.}

\item{group_by}{how to group your data, i.e the X axis of the plot - usually a metadata column in the seurat object.}

\item{cols}{optional: colours to fill the violins with.}

\item{pt.size}{size of plotted points.}

\item{sort}{sort X axis in descending order of average expression?}

\item{n_col}{if multiple features, how many columns to create in the final plot_grid object.}

\item{plot_hline}{plot a dashed horizontal line at 0?}
}
\value{
violin plot.
}
\description{
nicer looking version of the Seurat function VlnPlot.
}
