% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_GO_term.R
\name{plot_spatial_GO_term}
\alias{plot_spatial_GO_term}
\title{plot_spatial_GO_term}
\usage{
plot_spatial_GO_term(pathway_add, plot_tumor = T, plot_interface = F)
}
\arguments{
\item{pathway_add}{GO pathway you want to plot.}

\item{plot_tumor}{plot expression in tumor region?}

\item{plot_interface}{plot expression in interface region?}
}
\value{
plot.
}
\description{
add expression data for GO term to ST object metadata, and plot. Note: Seurat object for ABC must be loaded, and expression and pval matrices.
}
