% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_gene_Seurat.R
\name{search_for_gene_Seurat}
\alias{search_for_gene_Seurat}
\title{search_for_gene_Seurat}
\usage{
search_for_gene_Seurat(seurat_obj, gene, fixed = F)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{gene}{gene or portion of a gene name you want to search for}

\item{fixed}{if T, searches for exact matches only.}
}
\value{
matching gene names.
}
\description{
small function to search for presence of a gene name or string in the rownames of a Seurat object.
}
