% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlnplot_compare_ST_SC.R
\name{vlnplot_compare_ST_SC}
\alias{vlnplot_compare_ST_SC}
\title{vlnplot_compare_ST_SC}
\usage{
vlnplot_compare_ST_SC(
  SC_obj,
  ST_obj,
  features,
  pt_size = 0.3,
  sort = T,
  plot_hline = T
)
}
\arguments{
\item{SC_obj}{single cell Seurat object.}

\item{ST_obj}{ST Seurat object.}

\item{features}{gene(s) you want to plot expression of.}

\item{pt_size}{point size.}

\item{sort}{sort X-axis in descending order of mean expression?}

\item{plot_hline}{plot a dashed horizontal line at 0?}

\item{dims_plot}{dimensions to plot.}
}
\value{
violin plots.
}
\description{
nicer looking version of the Seurat function VlnPlot, to compare expression of gene(s) in scRNA-seq and ST datasets.
}
