% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_test_expr.R
\name{wilcox_test_expr}
\alias{wilcox_test_expr}
\title{wilcox_test_expr}
\usage{
wilcox_test_expr(seurat_obj, gene, group_by)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{gene}{gene to compare expression data across groups for.}

\item{group_by}{groups you want to compare, should be a column in metadata.}
}
\value{
P value matrix.
}
\description{
wilcoxon rank-sum test for multiple groups, to compare expression of a specific gene.
}
