% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_scatter_plot.R
\name{nice_scatter_plot}
\alias{nice_scatter_plot}
\title{nice_scatter_plot}
\usage{
nice_scatter_plot(
  df,
  x_var,
  y_var,
  fit_line = T,
  calculate_rsq = T,
  calculate_R = T
)
}
\arguments{
\item{df}{dataframe containing the data you want to plot.}

\item{x_var}{variable to be plotted on the x axis, as string.}

\item{y_var}{variable to be plotted on the y axis, as string.}

\item{fit_line}{if T, will fit a line using linear regression.}

\item{calculate_rsq}{if T, will calculate Rsq.}

\item{calculate_R}{if T, will calculate R.}
}
\value{
scatter plot. R and Rsq will be printed on plot if calculated.
}
\description{
create a nice looking scatter plot comparing 2 variables, with optional linear regression and calculation of R and Rsq.
}
